Introduce StepCardMenu
