+++ source/javascripts/components/unified-editor/WorkflowCard/components/StepCard.tsx
@@ -2,22 +2,7 @@ import { memo, ReactNode, useMemo, MouseEvent } from 'react';
 
 import { CSS } from '@dnd-kit/utilities';
 import { useSortable } from '@dnd-kit/sortable';
-import {
-  Avatar,
-  Box,
-  ButtonGroup,
-  Card,
-  CardProps,
-  ColorButton,
-  Divider,
-  Icon,
-  OverflowMenu,
-  OverflowMenuItem,
-  Skeleton,
-  SkeletonBox,
-  Text,
-  Tooltip,
-} from '@bitrise/bitkit';
+import { Avatar, Box, Card, CardProps, ColorButton, Icon, Skeleton, SkeletonBox, Text, Tooltip } from '@bitrise/bitkit';
 import { Popover, PopoverAnchor, PopoverArrow, PopoverBody, PopoverContent } from '@chakra-ui/react';
 import { useLocalStorage } from 'usehooks-ts';
 import useStep from '@/hooks/useStep';
@@ -27,9 +12,7 @@ import useDefaultStepLibrary from '@/hooks/useDefaultStepLibrary';
 import StepService from '@/core/services/StepService';
 import { Step } from '@/core/models/Step';
 import VersionUtils from '@/core/utils/VersionUtils';
-import useBitriseYmlStore from '@/hooks/useBitriseYmlStore';
-import generateUniqueEntityId from '@/core/utils/CommonUtils';
-import useFeatureFlag from '@/hooks/useFeatureFlag';
+import StepCardMenu from '@/components/unified-editor/WorkflowCard/components/StepCardMenu';
 import useReactFlowZoom from '../hooks/useReactFlowZoom';
 import { useSelection, useStepActions } from '../contexts/WorkflowCardContext';
 import { SortableStepItem } from '../WorkflowCard.types';
@@ -102,15 +85,10 @@ const StepCard = ({
     onCloneStepInStepBundle,
     onDeleteStep,
     onDeleteStepInStepBundle,
-    onGroupStepsToStepBundle,
     onSelectStep,
     onUpgradeStep,
     onUpgradeStepInStepBundle,
   } = useStepActions();
-
-  const enableStepBundles = useFeatureFlag('enable-wfe-step-bundles-ui');
-
-  const existingStepBundleIds = useBitriseYmlStore((s) => Object.keys(s.yml.step_bundles || {}));
 
   const {
     error,
@@ -148,10 +126,7 @@ const StepCard = ({
   const title = step?.title || step?.cvs || '';
   const isHighlighted = isSelected({ stepBundleId, stepIndex, workflowId });
   const { library } = StepService.parseStepCVS(step?.cvs || '', defaultStepLibrary);
-  const { isStep } = StepService;
-  const latestMajor = VersionUtils.latestMajor(step?.resolvedInfo?.versions)?.toString() ?? '';
 
-  const isSimpleStep = step && isStep(step.cvs, library);
   const isButton = !!onSelectStep;
   const isPlaceholder = sortable.isDragging;
   const isUpgradable =
@@ -212,138 +187,17 @@ const StepCard = ({
       return null;
     }
 
-    const suffix = selectedStepIndices && selectedStepIndices.length > 1 ? 's' : '';
-    const menuItems = [];
-    if (isUpgradable && (selectedStepIndices?.length === 1 || !isHighlighted)) {
-      menuItems.push(
-        <OverflowMenuItem
-          key="upgrade"
-          leftIconName="ArrowUp"
-          onClick={(e) => {
-            e.stopPropagation();
-            if (workflowId && onUpgradeStep) {
-              onUpgradeStep(workflowId, stepIndex, latestMajor);
-            }
-            if (stepBundleId && onUpgradeStepInStepBundle) {
-              onUpgradeStepInStepBundle(stepBundleId, stepIndex, latestMajor);
-            }
-          }}
-        >
-          Update Step version
-        </OverflowMenuItem>,
-      );
-    }
-    if ((enableStepBundles || !stepBundleId) && isSimpleStep) {
-      menuItems.push(
-        <OverflowMenuItem
-          key="group"
-          leftIconName="Steps"
-          onClick={(e) => {
-            e.stopPropagation();
-            if (onGroupStepsToStepBundle && selectedStepIndices) {
-              const generatedId = generateUniqueEntityId(existingStepBundleIds, 'Step_bundle');
-              const indices = isHighlighted ? selectedStepIndices : [stepIndex];
-              onGroupStepsToStepBundle(workflowId, stepBundleId, generatedId, indices);
-            }
-          }}
-        >
-          New bundle with {isHighlighted ? selectedStepIndices?.length : 1} Step
-          {suffix}
-        </OverflowMenuItem>,
-      );
-    }
-    if (selectedStepIndices?.length === 1 || !isHighlighted) {
-      menuItems.push(
-        <OverflowMenuItem
-          key="duplicate"
-          leftIconName="Duplicate"
-          onClick={(e) => {
-            e.stopPropagation();
-            if (workflowId && onCloneStep) {
-              onCloneStep(workflowId, stepIndex);
-            }
-            if (stepBundleId && onCloneStepInStepBundle) {
-              onCloneStepInStepBundle(stepBundleId, stepIndex);
-            }
-          }}
-        >
-          Duplicate Step
-        </OverflowMenuItem>,
-      );
-    }
-    if (menuItems.length > 0) {
-      menuItems.push(<Divider key="divider" my="8" />);
-    }
-    menuItems.push(
-      <OverflowMenuItem
-        isDanger
-        key="remove"
-        leftIconName="Trash"
-        onClick={(e) => {
-          e.stopPropagation();
-          if (workflowId && onDeleteStep) {
-            if (isHighlighted && selectedStepIndices) {
-              onDeleteStep(workflowId, selectedStepIndices);
-            } else {
-              onDeleteStep(workflowId, [stepIndex]);
-            }
-          }
-          if (stepBundleId && onDeleteStepInStepBundle) {
-            if (isHighlighted && selectedStepIndices) {
-              onDeleteStepInStepBundle(stepBundleId, selectedStepIndices);
-            } else {
-              onDeleteStepInStepBundle(stepBundleId, [stepIndex]);
-            }
-          }
-        }}
-      >
-        Delete Step{suffix}
-      </OverflowMenuItem>,
-    );
-
     return (
-      <ButtonGroup spacing="0" display="flex">
-        <OverflowMenu
-          placement="bottom-end"
-          size="md"
-          buttonSize="xs"
-          buttonProps={{
-            'aria-label': 'Show step actions',
-            iconName: 'MoreVertical',
-            onClick: (e) => {
-              e.stopPropagation();
-            },
-            display: 'none',
-            _groupHover: { display: 'inline-flex' },
-            _active: { display: 'inline-flex' },
-          }}
-        >
-          {...menuItems}
-        </OverflowMenu>
-      </ButtonGroup>
+      <StepCardMenu
+        isHighlighted={isHighlighted}
+        isUpgradable={isUpgradable}
+        step={step}
+        stepBundleId={stepBundleId}
+        stepIndex={stepIndex}
+        workflowId={workflowId}
+      />
     );
-  }, [
-    workflowId,
-    stepBundleId,
-    isDragging,
-    isUpgradable,
-    isClonable,
-    isRemovable,
-    selectedStepIndices,
-    isHighlighted,
-    enableStepBundles,
-    isSimpleStep,
-    onUpgradeStep,
-    onUpgradeStepInStepBundle,
-    stepIndex,
-    latestMajor,
-    onGroupStepsToStepBundle,
-    existingStepBundleIds,
-    onCloneStep,
-    onCloneStepInStepBundle,
-    onDeleteStep,
-    onDeleteStepInStepBundle,
-  ]);
+  }, [workflowId, stepBundleId, isDragging, isUpgradable, isClonable, isRemovable, isHighlighted, step, stepIndex]);
 
   return (
     <Card ref={sortable.setNodeRef} {...cardProps} style={style}>
